+++ src/main/java/com/revolvingmadness/sculk/language/builtins/classes/instances/MethodInstance.java
@@ -31,7 +31,7 @@ public class MethodInstance extends BuiltinMethod {
     public BuiltinClass call(Interpreter interpreter, List<BuiltinClass> arguments) {
         interpreter.variableTable.enterScope();
 
-        int maxArguments = Sculk.server.getGameRules().getInt(SculkGamerules.MAX_ARGUMENTS);
+        int maxArguments = Sculk.server.getGameRules().getInt(SculkGamerules.MAX_ARGUMENT_COUNT);
 
         if (arguments.size() > maxArguments) {
             throw new MaxArgumentError("Function '" + this.name + "' has more than " + maxArguments + " argument(s)");