+++ src/main/java/com/revolvingmadness/sculk/language/builtins/functions/AbsFunction.java
@@ -3,6 +3,10 @@ package com.revolvingmadness.sculk.language.builtins.functions;
 import com.revolvingmadness.sculk.language.ErrorHolder;
 import com.revolvingmadness.sculk.language.builtins.classes.BuiltinClass;
 import com.revolvingmadness.sculk.language.builtins.classes.BuiltinFunction;
+import com.revolvingmadness.sculk.language.builtins.classes.instances.FloatInstance;
+import com.revolvingmadness.sculk.language.builtins.classes.instances.IntegerInstance;
+import com.revolvingmadness.sculk.language.builtins.classes.types.FloatType;
+import com.revolvingmadness.sculk.language.builtins.classes.types.IntegerType;
 import com.revolvingmadness.sculk.language.interpreter.Interpreter;
 
 import java.util.List;
@@ -16,6 +20,14 @@ public class AbsFunction extends BuiltinFunction {
 
         BuiltinClass object = arguments.get(0);
 
-        return object.absoluteValue(object.getType());
+        if (object.instanceOf(new IntegerType())) {
+            return new IntegerInstance(Math.abs(object.toInteger()));
+        }
+
+        if (object.instanceOf(new FloatType())) {
+            return new FloatInstance(Math.abs(object.toFloat()));
+        }
+
+        throw ErrorHolder.argumentRequiresType(1, "abs", new IntegerType(), object.getType());
     }
 }